data/requirements.txt
pandas
mysql-connector-python
scikit-learn
matplotlib
schedule
python-dotenv
requests